[FIX] hr_timesheet_purchase_order: fields for recurrency moved to billing_partner
