+++ hr_timesheet_purchase_order/__manifest__.py
@@ -1,3 +1,6 @@
+# Copyright (C) 2024 Cetmix OÜ
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+
 {
     "name": "HR Timesheet Purchase Order",
     "version": "14.0.1.0.0",
@@ -17,6 +20,7 @@
         "data/ir_actions_server.xml",
         "data/hr_timesheet_cron.xml",
         "views/hr_employee_view.xml",
+        "views/res_partner_view.xml",
         "views/hr_timesheet_sheet_view.xml",
         "views/purchase_order_view.xml",
         "views/res_config_settings_view.xml",
