+++ hr_timesheet_purchase_order/models/__init__.py
@@ -1,6 +1,6 @@
+from . import res_partner
 from . import hr_timesheet_recurrence
 from . import hr_employee_base
-from . import hr_employee
 from . import hr_timesheet_sheet
 from . import res_config_settings
 from . import res_company
