--- hr_timesheet_purchase_order/models/hr_employee.py
@@ -1,7 +1,0 @@
-from odoo import fields, models
-
-
-class HR(models.Model):
-    _inherit = "hr.employee"
-
-    recurrence_id = fields.Many2one("hr.timesheet.recurrence", copy=False)
