+++ hr_timesheet_purchase_order/models/res_config_settings.py
@@ -1,3 +1,6 @@
+# Copyright (C) 2024 Cetmix OÜ
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+
 from odoo import fields, models
